Basis abi code generator (#7)
* swapping vector container with regular pointer

* making operator use generic container

* making seed state an array.

* changing  back to vector and adding lhss to bit_set

* change delcaration order

* redefining return type,

* fixing append and removing typedefs

* changing dit_integer_t to int.

* fixing typo in symmetry return args

* removing impl for conj and real

* adding complex_ops.h for abi

* adding code generator for abi

* adding optional functions for std::complex

* making flags for windows.

* swapping if

* removing abi from git tracking.

* fixing paranthesis

* removing unused variables

* making space/symmetry public members

* fixing return type for getting fock state

* adding two_body terms

* adding back case to bits switch

* adding check for diagonal matrix elements.

* adding fullspace definnitions exluding boost types

* adding boost compilation to include paths with option

* completed basis_abi design.

+++ src/quspin_core/includes/quspin/basis/types.h
@@ -15,7 +15,7 @@ typedef std::uint16_t uint16_t;
 typedef std::uint32_t uint32_t;
 typedef std::uint64_t uint64_t;
 
-typedef uint_fast8_t dit_integer_t; 
+typedef int dit_integer_t; 
 
 #ifdef USE_BOOST
 typedef boost::multiprecision::uint128_t uint128_t;
